moves ConfigClientImpl to internal package
