+++ src/main/java/cloud/prefab/client/PrefabCloudClient.java
@@ -1,5 +1,6 @@
 package cloud.prefab.client;
 
+import cloud.prefab.client.internal.ConfigClientImpl;
 import cloud.prefab.client.util.Cache;
 import cloud.prefab.client.util.NoopCache;
 import com.google.common.base.Preconditions;
