+++ src/test/java/cloud/prefab/client/config/logging/LogbackConfigListenerTest.java
@@ -5,7 +5,7 @@ import static org.assertj.core.api.Assertions.assertThat;
 import ch.qos.logback.classic.Level;
 import ch.qos.logback.classic.Logger;
 import ch.qos.logback.classic.LoggerContext;
-import cloud.prefab.client.ConfigClientImpl;
+import cloud.prefab.client.internal.ConfigClientImpl;
 import org.junit.jupiter.api.Test;
 import org.slf4j.LoggerFactory;
